% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trip_number.R
\name{add_trip_number}
\alias{add_trip_number}
\title{Add trip number to a series of locations using a base if the animal move or not outside of the buffer}
\usage{
add_trip_number(my_df, my_trip_number_sequence)
}
\arguments{
\item{my_df}{A data frame with central_location and num_sequence}

\item{my_trip_number_sequence}{a number to continue trip sequence}
}
\value{
vector with data
}
\description{
Add trip number to a series of locations using a base if the animal move or not outside of the buffer
}
