% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gaps.R
\name{calculate_gaps}
\alias{calculate_gaps}
\title{A function to calculate gaps between locations inside trips}
\usage{
calculate_gaps(
  GPS_data = GPS_data,
  column_datetime = column_datetime,
  column_tripnumber = column_tripnumber
)
}
\arguments{
\item{GPS_data}{the data containing date, time and trip number}

\item{column_datetime}{a column with date and time information already in POSIXct format}

\item{column_tripnumber}{a column with trip_number or any other separator}
}
\value{
the same data frame with an additional column called timedif_min
}
\description{
A function to calculate gaps between locations inside trips
}
\examples{
gull_gaps<-calculate_gaps(GPS_data = gull_data,column_datetime = 'DateTime',
column_tripnumber = 'trip_number')
}
