% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_leaving_returning.R
\name{calculate_leaving_returning}
\alias{calculate_leaving_returning}
\title{A function to calculate the distance from leaving and returning to the colony}
\usage{
calculate_leaving_returning(
  my_locs = my_locs,
  my_central_location = my_central_location
)
}
\arguments{
\item{my_locs}{a data frame with Longitude, Latitude,and trip_number}

\item{my_central_location}{a data frame with Longitude and Latitude}
}
\value{
a data frame with the trip number, the distance of leaving and returning, and the coordinates of the central location
}
\description{
A function to calculate the distance from leaving and returning to the colony
}
\examples{
calculate_leaving_returning(my_locs=Path_trips,my_central_location=data.frame(Longitude=-110.33,Latitude=24.287))
}
