% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_maxdist.R
\name{calculate_maxdist}
\alias{calculate_maxdist}
\title{Calculate maximum distance per trip}
\usage{
calculate_maxdist(
  my_data = my_data,
  central_location = central_location,
  divider = divider
)
}
\arguments{
\item{my_data}{a data frame with 'Longitude','Latitude'}

\item{central_location}{a data frame with 'Longitude','Latitude'}

\item{divider}{column name for divider}
}
\value{
returns data frame with maximum distance per trip
}
\description{
Calculate maximum distance per trip
}
\examples{
calculate_maxdist(my_data = GSM_trips, central_location = data.frame(Longitude=-110.34,Latitude=24.28),divider="trip_number")
}
