% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_params.R
\name{calculate_params}
\alias{calculate_params}
\title{Calculates trip duration using locations, and gap intervals per trip}
\usage{
calculate_params(
  my_locs = my_locs,
  my_daytime = my_daytime,
  my_format = my_format,
  my_units = my_units,
  my_separator = my_separator
)
}
\arguments{
\item{my_locs}{a data frame with all locations}

\item{my_daytime}{the name of the column where day and time information is provided}

\item{my_format}{the format in the column POSTIXct}

\item{my_units}{pick "hours", "minutes", "seconds".}

\item{my_separator}{a column to separate trips, for example 'trip_number'}
}
\value{
regresa un nuevo data frame con la información del viaje, cuando inicio, cuando termino y la duración
}
\description{
Calculates trip duration using locations, and gap intervals per trip
}
