% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length_criteria.R
\name{check_length_criteria}
\alias{check_length_criteria}
\title{This function is just to provide a baseline information from the foraging trips}
\usage{
check_length_criteria(my_params = my_params)
}
\arguments{
\item{my_params}{a data frame with trip parameters}
}
\value{
prints information about your parameters
}
\description{
This function is just to provide a baseline information from the foraging trips
}
