% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_locs.R
\name{classify_locs}
\alias{classify_locs}
\title{Add trip classification to the locations}
\usage{
classify_locs(Inter_params = Inter_params, Inter_locs = Inter_locs)
}
\arguments{
\item{my_locs}{a data frame with trip parameters}
}
\value{
returns four new columns with criteria to keep or recalculate trip parameters
}
\description{
Add trip classification to the locations
}
