% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_params.R
\name{classify_params}
\alias{classify_params}
\title{Add trip classification on the trip parameters}
\usage{
classify_params(my_params = my_params)
}
\arguments{
\item{my_params}{a data frame with trip parameters}
}
\value{
returns four new columns with criteria to keep or recalculate trip parameters
}
\description{
Add trip classification on the trip parameters
}
