% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_notrips.R
\name{compare_notrips}
\alias{compare_notrips}
\title{Compare number of trips from parameters and locations}
\usage{
compare_notrips(my_params = my_params, my_locs = my_locs)
}
\arguments{
\item{my_params}{the data frame with the parameters - must have trip number}

\item{my_locs}{the data frame with the locations - must have trip id}
}
\value{
information of which trips might be missing
}
\description{
Compare number of trips from parameters and locations
}
\examples{
compare_notrips(my_params=GSM_params,my_locs=GSM_trips)
}
