% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue_trip_sequence.R
\name{continue_trip_sequence}
\alias{continue_trip_sequence}
\title{A function to calculate gaps between locations inside trips}
\usage{
continue_trip_sequence(my_previous)
}
\arguments{
\item{my_previous}{a data frame containing information of trip_id}
}
\value{
the same data frame with an additional column called timedif_min
}
\description{
A function to calculate gaps between locations inside trips
}
\examples{
Previous_params<-as.data.frame(trip_id="trip_00405")
trip_number_sequence<-continue_trip_sequence(my_previous=Previous_params$trip_id)

}
