% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{Creates a buffer around a point}
\usage{
create_buffer(central_point = central_point, buffer_km = buffer_km)
}
\arguments{
\item{central_pont}{a data frame with latitude and longitude}
}
\value{
returns a buffer in spatial new_central_buffer
}
\description{
Creates a buffer around a point
}
