% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_per_trip.R
\name{distances_per_trip}
\alias{distances_per_trip}
\title{A function to calculate distance between points}
\usage{
distances_per_trip(my_df = my_df, my_divider = my_divider)
}
\arguments{
\item{my_df}{a data frame with a divider, latitude and longitude}

\item{my_divider}{the name of the column to divide}
}
\value{
a data frame with the distances between points
}
\description{
A function to calculate distance between points
}
\examples{
Path_distances<-distances_per_trip(my_df=Path_trips,my_divider='trip_number')
}
