% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_df_to_st.R
\name{from_df_to_st}
\alias{from_df_to_st}
\title{Converts a data frame into a sf simple features, uses 4326 as CRS}
\usage{
from_df_to_st(my_df)
}
\arguments{
\item{Reevaluate_locs}{a data frame with latitude and longitude}
}
\value{
returns a simple features, longitude and latitude are now in the geometry information
}
\description{
Converts a data frame into a sf simple features, uses 4326 as CRS
}
