% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSM_battery.R
\docType{data}
\name{GSM_battery}
\alias{GSM_battery}
\title{GSM battery information}
\format{
Columns are ID, daytime,season,month,battery_charge,gaps_min
}
\usage{
data(GSM_battery)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 28209 obs and 6 variables
}
\examples{
data(GSM_battery)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
