% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSM_locs.R
\docType{data}
\name{GSM_locs}
\alias{GSM_locs}
\title{GSM locations information}
\format{
Columns are num_seq,
ID, daytime,season,month,
Longitude,Latitude,
inside,central_base,
battery_charge,gaps_min
}
\usage{
data(GSM_locs)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 28209 obs and 11 variables
}
\examples{
data(GSM_locs)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
