% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSM_params.R
\docType{data}
\name{GSM_params}
\alias{GSM_params}
\title{GSM foraging trip parameters}
\format{
Columns are trip_id,trip_month_id,central_loc,trip_start,trip_end,min_gap,max_gap,duration
#'
}
\usage{
data(GSM_params)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 133 obs and 8 variables
}
\examples{
data(GSM_params)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
