% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gull_data.R
\docType{data}
\name{gull_data}
\alias{gull_data}
\title{Track from one gull}
\format{
Columns are Latitude, Longitude, DateTime, ID, at_sea, trip_number
}
\usage{
data(gull_data)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 1458 obs and 6 variables
}
\examples{
data(gull_data)
}
\references{
Lerma et al. (2023)
}
\keyword{datasets}
