% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_trips_reevaluation.R
\name{identify_trips_reevaluation}
\alias{identify_trips_reevaluation}
\title{Identifies trips from a data set}
\usage{
identify_trips_reevaluation(
  my_trip = my_trip,
  my_central_location = my_central_location,
  my_previous_params = my_previous_params
)
}
\arguments{
\item{my_trip}{trip locations}

\item{my_central_location}{data frame with longitude and latitude}

\item{my_previous_params}{a data frame containing trip id to keep the sequence, and not repetead trip numbers}
}
\value{
a data frame containing trip number
}
\description{
Identifies trips from a data set
}
\examples{
identify_trips_reevaluation(my_trip=Reevaluate_trips,my_central_location=data.frame(Longitude=-110.33,Latitude=24.15),my_previous_params=Params_criteria$trip_id)
}
