% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_trips.R
\name{interpolate_trips}
\alias{interpolate_trips}
\title{Interpolate_trips, interpolate your tracking data to a specific interval}
\usage{
interpolate_trips(
  my_df = my_df,
  interval = interval,
  column_datetime = column_datetime,
  datetime_format = datetime_format,
  column_lat = column_lat,
  column_lon = column_lon,
  column_trip = column_trip
)
}
\arguments{
\item{my_df}{is the data frame to interpolate}

\item{interval}{is the interval you want to use, in seconds, for example 900 s is 15 minutes}

\item{datetime_format}{is the format you use for your date and time, for example "\%d/\%m/\%Y \%H:\%M:\%S", "\%d-\%m-\%Y \%H:\%M:\%S" or "\%Y-\%m-\%d \%H:\%M:\%S"}

\item{column_lat}{is the name of the column where you have the Latitude}

\item{column_lon}{is the name of the column where you have the Longitude}

\item{column_trip}{is the name of the column where you have the Trip number, if you dont have number of trip please check the function count_trips}

\item{column_date}{is the name of the column where you have the Date and Time}
}
\value{
Gives back a data frame with the interpolated data, only four columns will be returned: dt, Longitude, Latitude, and trip_number
}
\description{
Interpolate_trips, interpolate your tracking data to a specific interval
}
\examples{
interpolate_trips(my_df=Interpolation_trips,interval='900 sec',column_datetime='daytime',column_trip='trip_number',column_lat='Latitude',column_lon='Longitude',datetime_format="\%Y-\%m-\%d \%H:\%M:\%S")
}
