% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation_params.R
\docType{data}
\name{Interpolation_params}
\alias{Interpolation_params}
\title{GSM foraging trip parameters information candidates for interpolation}
\format{
Columns are
trip_id,trip_month_id,central_loc,
trip_start,trip_end,
min_gap,max_gap,
duration,
trip_size,resolution,params_analyses,interpolation
}
\usage{
data(Interpolation_params)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 22 obs and 12 variables
}
\examples{
data(Interpolation_params)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
