% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation_trips.R
\docType{data}
\name{Interpolation_trips}
\alias{Interpolation_trips}
\title{GSM locations information for interpolation}
\format{
Columns are
ID,trip_number,num_seq,
daytime,season,month,
Longitude,Latitude,
inside,central_base,
battery_charge,
gaps_min,
trip_size,params_analyses,resolution,interpolation
}
\usage{
data(Interpolation_trips)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 376 obs and 16 variables
}
\examples{
data(Interpolation_trips)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
