% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over.R
\name{over}
\alias{over}
\title{Gives information if the polygon intersects the locations or not}
\usage{
over(my_locations = my_locations, my_polygon = my_polygon)
}
\arguments{
\item{my_locations}{s Simple feature with a data frame}

\item{my_polygon}{a Simple feature collection with 1 feature and 0 fields}
}
\value{
vector with data
}
\description{
Gives information if the polygon intersects the locations or not
}
