% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path_trips.R
\docType{data}
\name{Path_trips}
\alias{Path_trips}
\title{GSM trips information}
\format{
Columns are ID, trip_number,num_seq,
daytime,season,month,
Longitude,Latitude,
inside,central_base,
battery_charge,gaps_min
}
\usage{
data(Path_trips)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 1588 obs and 12 variables
}
\examples{
data(Path_trips)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
