% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_check.R
\name{plot_check}
\alias{plot_check}
\title{Creates a plot to check locations}
\usage{
plot_check(my_locs = my_locs, my_central_location = my_central_location)
}
\arguments{
\item{my_locs}{a data frame with locations}

\item{my_central_location}{a data frame with the central location}
}
\value{
plot to evaluate if the central location is correct
}
\description{
Creates a plot to check locations
}
\examples{
plot_check(my_locs=GSM_locs,my_central_location = data.frame(Longitude=-110.34,Latitude=24.28))
}
