% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reevaluate_polygon.R
\docType{data}
\name{Reevaluate_polygon}
\alias{Reevaluate_polygon}
\title{Polygon of a buffer created from a central location}
\format{
contains geometry of sfc_POLYGON
}
\usage{
data(Reevaluate_polygon)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 1 geometry polygon
}
\examples{
data(Reevaluate_polygon)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
