% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reevaluate_sf.R
\docType{data}
\name{Reevaluate_sf}
\alias{Reevaluate_sf}
\title{GSM locations information a outlier trip}
\format{
Columns are ID,trip_number,daytime, gaps_min,num_seq,
season,month,
inside,central_base,
battery_charge,
geometry
}
\usage{
data(Reevaluate_sf)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 738 obs and 11 variables
}
\examples{
data(Reevaluate_sf)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
