% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reevaluate_trips.R
\docType{data}
\name{Reevaluate_trips}
\alias{Reevaluate_trips}
\title{GSM locations information only trip outlier}
\format{
Columns are ID,trip_number,daytime, gaps_min,num_seq,
season,month,
Longitude,Latitude,
inside,central_base,
battery_charge, central_location
}
\usage{
data(Reevaluate_trips)
}
\description{
Tracking data from one Yellow-footed gull tracked at La Paz, Mexico
Includes 738 obs and 13 variables
}
\examples{
data(Reevaluate_trips)
}
\references{
Lerma et al. (2025)
}
\keyword{datasets}
