% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_reevaluation.R
\name{subset_reevaluation}
\alias{subset_reevaluation}
\title{A function to subset a specific trip from the GSM data}
\usage{
subset_reevaluation(
  my_tripid = my_tripid,
  my_trip = my_trip,
  new_central_location = new_central_location,
  old_central_location = old_central_location
)
}
\arguments{
\item{my_tripid}{the trip to be subset}

\item{my_trip}{the data frame with the locations it should contain trip_number}

\item{new_central_location}{data frame with latitude and longitude of new central location}

\item{old_central_location}{data frame with latitude and longitude of old central location}
}
\value{
a data frame filtered
}
\description{
A function to subset a specific trip from the GSM data
}
\examples{
subset_reevaluation(my_trip=GSM_trips,my_tripid='trip_00528',new_central_location=data.frame(Longitude=-110.32,Latitude=24.15), old_central_location=data.frame(Longitude=-110.34,Latitude=24.28))
}
