% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gaps.R
\name{calculate_gaps}
\alias{calculate_gaps}
\title{A function to calculate gaps between locations inside trips}
\usage{
calculate_gaps(
  my_locs = my_locs,
  my_datetime = my_datetime,
  my_separator = my_separator
)
}
\arguments{
\item{my_locs}{the data containing date, time and trip number}

\item{my_datetime}{a column with date and time information already in POSIXct format}

\item{my_separator}{a column with trip_number or any other separator}
}
\value{
the same data frame with an additional column called timedif_min
}
\description{
A function to calculate gaps between locations inside trips
}
\examples{
calculate_gaps(my_locs = GSM_locs,my_datetime = 'daytime',my_separator = 'ID')
}
